build: make_* → build_*, one UT script, profile catalog everywhere, libemlog/-dev deb split (v1.3.0)
- All make_* scripts renamed to build_*; make_UTs_priv/pub/all merged into a
  single build_UTs.sh that compiles both suites as separate binaries (they
  must be — white-box vs black-box) and runs them in one run, with
  --build-only/--run-only to split the phases (also on build_UTs_release.sh
  and build_ITs.sh).
- Every test script now sources gcc_build_profiles.sh instead of carrying
  ad-hoc flag literals: coverage UTs = debug profile + coverage layer (-O0
  appended for exact gcov attribution), release UTs and IT = release profile.
- Honest per-suite coverage: gcovr gets each suite's build dir as a SEARCH
  PATH; --object-directory swept the whole tree and silently merged both
  suites into every "per-suite" number (private really covers 12.7%, not
  the reported 96.4%). Combined report unchanged (96.4% line / 82.7% branch).
- build_deb.sh now produces the standard Debian library split the releases
  were supposed to publish all along: libemlog (runtime .so + ldconfig
  hooks) and libemlog-dev (header, .a, linker symlink, Depends: exact
  version) — quality.yml installs both before the smoke test.
- emlog.c: _copy_cached_ts composes in a worst-case-sized local buffer;
  -Wformat-truncation=2 treats an unknown-size destination as size 1, so
  formatting into the caller's buffer could never be warning-clean. Audit and
  sanitize profiles now compile silent.
- integration_test.c: explicit casts silence -Wconversion under the release
  profile's strict warnings.
- README documents every script, flag, artifact, and package; RELEASING.md
  lists the real release assets; badges track master.

+++ tests/ITs/integration_test.c
@@ -100,7 +100,7 @@ int main(void)
         pthread_join(th[i], NULL);
 
     clock_gettime(CLOCK_MONOTONIC, &t1);
-    double elapsed = (t1.tv_sec - t0.tv_sec) + (t1.tv_nsec - t0.tv_nsec) / 1e9;
+    double elapsed = (double)(t1.tv_sec - t0.tv_sec) + (double)(t1.tv_nsec - t0.tv_nsec) / 1e9;
 
     const unsigned long expected = (unsigned long)nthreads * (unsigned long)msgs;
     const unsigned long seen     = atomic_load_explicit(&g_lines, memory_order_relaxed);
